Teacher Module

+++ src/services/api/manager/delete-manager.ts
@@ -17,6 +17,6 @@ export type TDeleteManagerResponse = unknown;
 // FUNCTION
 
 export const deleteManager = async ({ paths, params }: TDeleteManagerMaterials): Promise<TDeleteManagerResponse> => {
-  const response = await ApiService.delete(`/v1/${paths?.id}`, { params });
+  const response = await ApiService.delete(`/v1/api/admin/managers/${paths?.id}`, { params });
   return response.data;
 };
